Reviews, CRUD

+++ controllers/courses.js
@@ -48,6 +48,7 @@ exports.getCourse = asyncHandler(async (req, res , next)=>{
 // @access    Private
 exports.addCourse = asyncHandler(async (req, res, next) => {
   req.body.bootcamp = req.params.bootcampId;
+  req.body.user = req.user.id;
 
   const bootcamp = await Bootcamp.findById(req.params.bootcampId);
 
@@ -60,6 +61,16 @@ exports.addCourse = asyncHandler(async (req, res, next) => {
     );
   }
 
+  // Make sure user is bootcamp owner
+  if (bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin') {
+    return next(
+      new ErrorResponse(
+        `User ${req.user.id} is not authorized to add a course to bootcamp ${bootcamp._id}`,
+        401
+      )
+    );
+  }
+
   const course = await Course.create(req.body);
 
   res.status(200).json({
@@ -67,6 +78,7 @@ exports.addCourse = asyncHandler(async (req, res, next) => {
     data: course
   });
 });
+
 
 // @desc      Update course
 // @route     PUT /api/v1/bootcamps/:bootcampId/courses
@@ -82,6 +94,17 @@ exports.updateCourse = asyncHandler(async (req, res, next) => {
       )
     );
   }
+
+   // Make sure user is course owner
+   if (course.user !== req.user.id && req.user.role !== 'admin') {
+    return next(
+      new ErrorResponse(
+        `User ${req.user.id} is not authorized to update course ${course._id}`,
+        401
+      )
+    );
+  }
+
 
 course = await Course.findByIdAndUpdate(req.params.id, req.body, {
   new: true,
@@ -109,6 +132,16 @@ exports.deleteCourse = asyncHandler(async (req, res, next) => {
       )
     );
   }
+
+// Make sure user is course owner
+if (course.user !== req.user.id && req.user.role !== 'admin') {
+  return next(
+    new ErrorResponse(
+      `User ${req.user.id} is not authorized to delete course ${course._id}`,
+      401
+    )
+  );
+}
 
 await course.deleteOne()
